+++ Pandora/Paradigm/Controlflow/Joint/Adaptable.hs
@@ -10,57 +10,65 @@ import Pandora.Pattern.Functor.Monad (Monad)
 import Pandora.Paradigm.Controlflow.Joint.Schematic (Schematic)
 import Pandora.Paradigm.Controlflow.Joint.Transformer.Monadic (Monadic (lay, wrap), (:>))
 
-class Adaptable eff schema where
+class Covariant u => Adaptable t u where
 	{-# MINIMAL adapt #-}
-	adapt :: eff ~> schema
+	adapt :: t ~> u
 
 type Layable t u = (Monadic t, Covariant u)
 type Wrappable t u = (Monadic t, Pointable u)
 
-instance Adaptable t t where
+instance Covariant t => Adaptable t t where
 	adapt = identity
 
-instance Layable t u => Adaptable u (t :> u) where
+instance (Covariant (t :> u), Layable t u) => Adaptable u (t :> u) where
 	adapt = lay
 
-instance Wrappable t u => Adaptable t (t :> u) where
+instance (Covariant (t :> u), Wrappable t u) => Adaptable t (t :> u) where
 	adapt = wrap
 
 instance
-	( Layable t (Schematic Monad u v)
+	( Covariant (t :> u :> v)
+	, Layable t (Schematic Monad u v)
 	, Wrappable u v
 	) => Adaptable u (t :> u :> v) where
 	adapt = lay . wrap
 
 instance
-	( Layable t (Schematic Monad u v)
+	( Covariant (t :> u :> v)
+	, Layable t (Schematic Monad u v)
 	, Layable u v
 	) => Adaptable v (t :> u :> v) where
 	adapt = lay . lay
 
 instance
-	( Layable t (Schematic Monad u (v :> w))
+	( Covariant (t :> u :> v :> w)
+	, Layable t (Schematic Monad u (v :> w))
 	, Layable u (Schematic Monad v w)
 	, Wrappable v w
 	) => Adaptable v (t :> u :> v :> w) where
 	adapt = lay . lay . wrap
 
 instance
-	( Layable t (Schematic Monad u v)
+	( Covariant (t :> u :> v :> w)
+	, Layable t (Schematic Monad u v)
 	, Layable t (Schematic Monad u (v :> w))
 	, Layable u (Schematic Monad v w)
 	, Layable v w
 	) => Adaptable w (t :> u :> v :> w) where
 	adapt = lay . lay . lay
 
-instance (Layable t (Schematic Monad u (v :> w :> x))
+instance
+	( Covariant (t :> u :> v :> w :> x)
+	, Layable t (Schematic Monad u (v :> w :> x))
 	, Layable u (Schematic Monad v (w :> x))
 	, Layable v (Schematic Monad w x)
 	, Layable w x
 	) => Adaptable x (t :> u :> v :> w :> x) where
 	adapt = lay . lay . lay . lay
 
-instance (Layable t (Schematic Monad u (v :> w :> x))
+instance
+	( Covariant (t :> u :> v :> w :> x)
+	, Layable t (Schematic Monad u (v :> w :> x))
 	, Layable u (Schematic Monad v (w :> x))
 	, Layable v (Schematic Monad w x)
 	, Wrappable w x
@@ -68,7 +76,8 @@ instance (Layable t (Schematic Monad u (v :> w :> x))
 	adapt = lay . lay . lay . wrap
 
 instance
-	( Layable t (Schematic Monad u (v :> w :> x :> y))
+	( Covariant (t :> u :> v :> w :> x :> y)
+	, Layable t (Schematic Monad u (v :> w :> x :> y))
 	, Layable u (Schematic Monad v (w :> x :> y))
 	, Layable v (Schematic Monad w (x :> y))
 	, Layable w (Schematic Monad x y)
@@ -77,7 +86,8 @@ instance
 	adapt = lay . lay . lay . lay . lay
 
 instance
-	( Layable t (Schematic Monad u (v :> w :> x :> y))
+	( Covariant (t :> u :> v :> w :> x :> y)
+	, Layable t (Schematic Monad u (v :> w :> x :> y))
 	, Layable u (Schematic Monad v (w :> x :> y))
 	, Layable v (Schematic Monad w (x :> y))
 	, Layable w (Schematic Monad x y)
@@ -86,7 +96,8 @@ instance
 	adapt = lay . lay . lay . lay . wrap
 
 instance
-	( Layable t (Schematic Monad u (v :> w :> x :> y :> z))
+	( Covariant (t :> u :> v :> w :> x :> y :> z)
+	, Layable t (Schematic Monad u (v :> w :> x :> y :> z))
 	, Layable u (Schematic Monad v (w :> x :> y :> z))
 	, Layable v (Schematic Monad w (x :> y :> z))
 	, Layable w (Schematic Monad x (y :> z))
@@ -96,7 +107,8 @@ instance
 	adapt = lay . lay . lay . lay . lay . lay
 
 instance
-	( Layable t (Schematic Monad u (v :> w :> x :> y :> z))
+	( Covariant (t :> u :> v :> w :> x :> y :> z)
+	, Layable t (Schematic Monad u (v :> w :> x :> y :> z))
 	, Layable u (Schematic Monad v (w :> x :> y :> z))
 	, Layable v (Schematic Monad w (x :> y :> z))
 	, Layable w (Schematic Monad x (y :> z))
@@ -106,7 +118,8 @@ instance
 	adapt = lay . lay . lay . lay . lay . wrap
 
 instance
-	( Layable t (Schematic Monad u (v :> w :> x :> y :> z :> f))
+	( Covariant (t :> u :> v :> w :> x :> y :> z :> f)
+	, Layable t (Schematic Monad u (v :> w :> x :> y :> z :> f))
 	, Layable u (Schematic Monad v (w :> x :> y :> z :> f))
 	, Layable v (Schematic Monad w (x :> y :> z :> f))
 	, Layable w (Schematic Monad x (y :> z :> f))
@@ -117,7 +130,8 @@ instance
 	adapt = lay . lay . lay . lay . lay . lay . lay
 
 instance
-	( Layable t (Schematic Monad u (v :> w :> x :> y :> z :> f))
+	( Covariant (t :> u :> v :> w :> x :> y :> z :> f)
+	, Layable t (Schematic Monad u (v :> w :> x :> y :> z :> f))
 	, Layable u (Schematic Monad v (w :> x :> y :> z :> f))
 	, Layable v (Schematic Monad w (x :> y :> z :> f))
 	, Layable w (Schematic Monad x (y :> z :> f))
@@ -128,7 +142,8 @@ instance
 	adapt = lay . lay . lay . lay . lay . lay . wrap
 
 instance
-	( Layable t (Schematic Monad u (v :> w :> x :> y :> z :> f :> h))
+	( Covariant (t :> u :> v :> w :> x :> y :> z :> f :> h)
+	, Layable t (Schematic Monad u (v :> w :> x :> y :> z :> f :> h))
 	, Layable u (Schematic Monad v (w :> x :> y :> z :> f :> h))
 	, Layable v (Schematic Monad w (x :> y :> z :> f :> h))
 	, Layable w (Schematic Monad x (y :> z :> f :> h))
@@ -140,7 +155,8 @@ instance
 	adapt = lay . lay . lay . lay . lay . lay . lay . lay
 
 instance
-	( Layable t (Schematic Monad u (v :> w :> x :> y :> z :> f :> h))
+	( Covariant (t :> u :> v :> w :> x :> y :> z :> f :> h)
+	, Layable t (Schematic Monad u (v :> w :> x :> y :> z :> f :> h))
 	, Layable u (Schematic Monad v (w :> x :> y :> z :> f :> h))
 	, Layable v (Schematic Monad w (x :> y :> z :> f :> h))
 	, Layable w (Schematic Monad x (y :> z :> f :> h))